renamings: terminal->top (everywhere incl code); threads-># (home)
